tarot1.6.py
My current iteration of the program. I have simplified the card drawing system. Next to randomize the 3 card lists tarot0-10&21 to increase variation of deck. Then to make a draw that's larger or smaller then 3 cards. then a feture to save passed draws for comperison and reference, finally front end design and imagery.